chore: move playground app to root folder
